Add files via upload
Dungeons and Dragons style dice roller that uses regular expressions and the random module to simulate dice rolling. Syntax is XdN+Y where X = the number of die to roll, N=the number of sides for the die, and Y is an optional modifier to add or subtract values from what is rolled. For example, 6d10+8 will roll 6 10-sided die and add 8 to the result. The bot responds in English.